Add files via upload
This is my first clone project in index.html in this project i used, CSS for styling and Script.js for functionality. with the help of ChatGPT, I was able to complete it . please give me feedback on my project- it will helps me grow and improve more affectively . I am still a beginner in the tech field, but i am trying to build projects by myself and learn step by step.